adapt update_user_interests parameter
